+++ R/backend-snowflake.R
@@ -42,11 +42,18 @@ sql_translation.Snowflake <- function(con) {
       # str_detect does not.  Left- and right-pad `pattern` with .* to get
       # str_detect-like behavior
       str_detect = function(string, pattern, negate = FALSE) {
-        if (isTRUE(negate)) {
-          sql_expr(!(((!!string)) %REGEXP% (".*" || (!!pattern) || ".*")))
-        } else {
-          sql_expr(((!!string)) %REGEXP% (".*" || (!!pattern) || ".*"))
-        }
+        sql_str_pattern_switch(
+          string = string,
+          pattern = {{ pattern }},
+          negate = function(string, pattern, negate = FALSE) {
+            if (isTRUE(negate)) {
+              sql_expr(!(((!!string)) %REGEXP% (".*" || (!!pattern) || ".*")))
+            } else {
+              sql_expr(((!!string)) %REGEXP% (".*" || (!!pattern) || ".*"))
+            }
+          },
+          f_fixed = sql_str_detect_fixed_instr("detect")
+        )
       },
       # On Snowflake, REGEXP_REPLACE is used like this:
       # REGEXP_REPLACE( <subject> , <pattern> [ , <replacement> ,